+++ vasttools/pipeline.py
@@ -62,6 +62,12 @@ class PipelineDirectoryError(Exception):
     """
     pass
 
+class MeasPairsDoNotExistError(Exception):
+    """
+    An error to indicate that the measurement pairs do not exist for a run.
+    """
+    pass
+
 
 class PipeRun(object):
     """
@@ -161,7 +167,22 @@ class PipeRun(object):
 
         self.logger = logging.getLogger('vasttools.pipeline.PipeRun')
         self.logger.debug('Created PipeRun instance')
+        
+        self._measurement_pairs_exists = self._check_measurement_pairs_file()
 
+    def _check_measurement_pairs_file(self):
+        measurement_pairs_exists = True
+        
+        for filepath in self.measurement_pairs_file:
+            if not os.path.isfile(filepath):
+                self.logger.warning(f"Measurement pairs file ({filepath}) does"
+                                    f" not exist. You will be unable to access"
+                                    f" measurement pairs or two-epoch metrics."
+                                    )
+                measurement_pairs_exists = False
+
+        return measurement_pairs_exists
+            
     def combine_with_run(
         self, other_PipeRun, new_name: Optional[str] = None
     ):
@@ -224,8 +245,23 @@ class PipeRun(object):
 
         # need to keep access to all the different pairs files
         # for two epoch metrics.
-        for i in other_PipeRun.measurement_pairs_file:
-            self.measurement_pairs_file.append(i)
+        orig_run_pairs_exist = self._measurement_pairs_exists
+        other_run_pairs_exist = other_PipeRun._measurement_pairs_exists
+
+        if orig_run_pairs_exist and other_run_pairs_exist:
+            for i in other_PipeRun.measurement_pairs_file:
+                self.measurement_pairs_file.append(i)
+        
+        elif orig_run_pairs_exist:
+            self.logger.warning("Not combining measurement pairs because they "
+                                " do not exist for the new run."
+                                )
+            self._measurement_pairs_exists = False
+
+        elif other_run_pairs_exist:
+            self.logger.warning("Not combining measurement pairs because they "
+                                " do not exist for the original run."
+                                )
 
         del sources_to_add
 
@@ -407,6 +443,13 @@ class PipeRun(object):
 
         return thesource
 
+    def _raise_if_no_pairs(self):
+        if not self._measurement_pairs_exists:
+             raise MeasPairsDoNotExistError("This method cannot be used as "
+                                            "the measurement pairs are not "
+                                            "available for this pipeline run."
+                                            )
+
     def load_two_epoch_metrics(self) -> None:
         """
         Loads the two epoch metrics dataframe, usually stored as either
@@ -421,7 +464,14 @@ class PipeRun(object):
 
         Returns:
             None
+
+        Raises:
+            MeasPairsDoNotExistError: The measurement pairs file(s) do not
+                exist for this run
         """
+        
+        self._raise_if_no_pairs()
+
         image_ids = self.images.sort_values(by='datetime').index.tolist()
 
         pairs_df = pd.DataFrame.from_dict(
@@ -890,6 +940,7 @@ class PipeAnalysis(PipeRun):
         Returns:
             The filtered measurement pairs dataframe.
         """
+
         if not self._loaded_two_epoch_metrics:
             self.load_two_epoch_metrics()
 
@@ -1041,7 +1092,18 @@ class PipeAnalysis(PipeRun):
         Returns:
             The regenerated sources_df.  A `pandas.core.frame.DataFrame`
             instance.
+        
+        Raises:
+            MeasPairsDoNotExistError: The measurement pairs file(s) do not
+                exist for this run
         """
+        
+        self._raise_if_no_pairs()
+
+        # Two epoch metrics
+        if not self._loaded_two_epoch_metrics:
+            self.load_two_epoch_metrics()
+
         if not self._vaex_meas:
             measurements_df = vaex.from_pandas(measurements_df)
 
@@ -1154,10 +1216,6 @@ class PipeAnalysis(PipeRun):
         sources_df = sources_df.join(
             self.sources[['new', 'new_high_sigma']],
         )
-
-        # Two epoch metrics
-        if not self._loaded_two_epoch_metrics:
-            self.load_two_epoch_metrics()
 
         if measurement_pairs_df is None:
             measurement_pairs_df = self._filter_meas_pairs_df(
@@ -1646,7 +1704,12 @@ class PipeAnalysis(PipeRun):
             Exception: 'plot_type' is not recognised.
             Exception: `plot_style` is not recognised.
             Exception: Pair with entered ID does not exist.
+            MeasPairsDoNotExistError: The measurement pairs file(s) do not
+                exist for this run
         """
+
+        self._raise_if_no_pairs()
+
         if not self._loaded_two_epoch_metrics:
             raise Exception(
                 "The two epoch metrics must first be loaded to use the"
@@ -1737,7 +1800,12 @@ class PipeAnalysis(PipeRun):
         Raises:
             Exception: The two epoch metrics must be loaded before using this
                 function.
+            MeasPairsDoNotExistError: The measurement pairs file(s) do not
+                exist for this run
         """
+        
+        self._raise_if_no_pairs()
+
         if not self._loaded_two_epoch_metrics:
             raise Exception(
                 "The two epoch metrics must first be loaded to use the"